pyproject.toml
[tool.poetry]
name = "fastapi-application"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
uvicorn= {extras = ["standard"], version = "^0.20.0"}
python-multipart = "^0.0.6"
pymongo = {extras = ["srv"], version = "^4.3.3"}
orjson = "^3.9.1"
boto3 = "^1.26.153"
pandas = "^2.0.2"
pypdf2 = "^3.0.1"
langchain = "^0.0.229"
openai = "^0.27.8"
tiktoken = "^0.4.0"
fastapi = "^0.100.0"
pydantic = "1.10.8"
pinecone-client = "^2.2.2"


[tool.poetry.group.dev.dependencies]
types-python-jose = "^3.3.4.7"
ruff = "^0.0.271"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
